+++ packages/gamut-labs/src/LayoutMenu/LayoutMenu.tsx
@@ -1,9 +1,8 @@
-import { Box, Logo, StrokeButton } from '@codecademy/gamut';
+import { Box, Flyout, Logo, StrokeButton } from '@codecademy/gamut';
 import { variant } from '@codecademy/gamut-styles';
 import styled from '@emotion/styled';
 import React, { useState } from 'react';
 
-import { Flyout } from '../Flyout';
 import { AccordionMenu, Section, SectionItem } from './AccordionMenu';
 import { LayoutMenuSection } from './LayoutMenuSection';
