+++ packages/gamut-labs/src/index.ts
@@ -1,50 +1,50 @@
-export * from './LandingPage';
-export * from './lib/breakpointHooks';
-export * from './Avatar';
-export * from './AvatarChooser';
-export * from './Byline';
-export * from './Breadcrumbs';
-export * from './CurriculumCard';
-export * from './CurriculumCard/Difficulty';
-export * from './CurriculumCard/Difficulty/types';
-export * from './Drawer';
-export * from './EditorialImage';
-export * from './ErrorContents';
-export * from './Flyout';
-export * from './Header/HeaderContainer';
-export * from './Header/HeaderTab';
-export * from './LayoutMenu';
-export * from './Loading';
-export * from './LogoFromSkillsoft';
-export * from './Logo';
-export * from './ProLogo';
-export * from './ProLogoAlt';
-export * from './TestimonialDeprecated';
-export * from './Testimonial';
-export * from './Coachmark';
-export * from './DropdownButton';
-export * from './HubCard';
-export * from './Popover';
+export * from './AppBar'; //
+export * from './AppBar/AppBarSection'; //
 export * from './AppHeader';
 export * from './AppHeader/AppHeaderElements/AppHeaderDropdown';
-export * from './AppHeader/AppHeaderElements/types';
 export * from './AppHeader/AppHeaderElements/AppHeaderLink';
 export * from './AppHeader/AppHeaderElements/AppHeaderLogo';
-export type { AppHeaderNotifications } from './Notifications/types';
-export type { AppHeaderSearch } from './AppHeader/Search/useHeaderSearch';
+export * from './AppHeader/AppHeaderElements/types';
 export * from './AppHeaderMobile';
 export * from './AppHeaderMobile/AppHeaderMainMenuMobile';
 export * from './AppHeaderMobile/AppHeaderSubMenuMobile';
+export * from './Avatar'; //
+export * from './AvatarChooser'; //
+export * from './Banner'; //
+export * from './BetaSticker'; //
+export * from './Byline'; //
+export * from './CurriculumCard'; //
+export * from './CurriculumCard/Difficulty'; //
+export * from './CurriculumCard/Difficulty/types'; //
+export * from './DropdownButton'; //
+export * from './EditorialImage'; //
+export * from './EmptySection'; //
+export * from './ErrorContents'; //
 export * from './GlobalFooter';
 export * from './GlobalHeader';
-export type { AnonHeaderVariant } from './GlobalHeader/types';
 export * from './GlobalPage';
-export * from './PageSection';
+export * from './Header/HeaderContainer';
+export * from './Header/HeaderTab';
+export * from './HorizontalScrollMenu'; //
+export * from './HubCard'; //
+export * from './InfoCard';
+export * from './Interstitial';
+export * from './LandingPage';
+export * from './LayoutMenu';
+export * from './lib/breakpointHooks';
 export * from './ListSection';
-export * from './EmptySection';
+export * from './Loading';
+export * from './Logo';
+export * from './LogoFromSkillsoft';
+export * from './PageAlerts';
+export * from './PageSection';
+export * from './PausableImage';
+export * from './ProLogo';
+export * from './ProLogoAlt';
 export * from './SocialMediaSharing';
 export * from './TabsExperimental';
-export * from './PageAlerts';
-export * from './InfoCard';
-export * from './PausableImage';
-export * from './HorizontalScrollMenu';
+export * from './Testimonial';
+export * from './TestimonialDeprecated';
+export type { AnonHeaderVariant } from './GlobalHeader/types';
+export type { AppHeaderNotifications } from './Notifications/types';
+export type { AppHeaderSearch } from './AppHeader/Search/useHeaderSearch';
