+++ packages/gamut/src/Breadcrumbs/index.tsx
@@ -1,7 +1,9 @@
-import { Anchor, Box, FlexBox, Text } from '@codecademy/gamut';
 import { css } from '@codecademy/gamut-styles';
 import styled from '@emotion/styled';
 import React from 'react';
+
+import { Anchor } from '../Anchor';
+import { Box, FlexBox } from '../Box';
 
 const BreadcrumbAnchor = styled(Anchor)(
   css({
