+++ packages/gamut/src/index.tsx
@@ -1,65 +1,65 @@
-export * from './Anchor';
-export * from './Alert';
 export * from './Accordion';
 export * from './AccordionArea';
 export * from './AccordionButton';
+export * from './Alert';
+export * from './Anchor';
 export * from './Animation';
-export * from './AppBar';
-export * from './AppBar/AppBarSection';
 export * from './AppWrapper';
 export * from './Badge';
-export * from './Banner';
+export * from './BodyPortal';
 export * from './Box';
+export * from './Breadcrumbs';
 export * from './Button';
 export type { ButtonBaseElements } from './ButtonBase/ButtonBase';
 export * from './ButtonDeprecated';
 export * from './ButtonDeprecatedBase';
-export * from './BodyPortal';
 export * from './Card';
+export * from './Coachmark';
 export * from './ConnectedForm';
-export * from './DeprecatedCard/CardShell';
 export * from './ContentContainer';
+export * from './DataList';
+export * from './DeprecatedCard/CardShell';
+export * from './Drawer';
 export * from './FlexBox';
+export * from './FloatingCard/FloatingCard';
+export * from './Flyout';
 export * from './FocusTrap';
 export * from './Form';
 export * from './GridForm';
 export * from './HiddenText';
-export * from './Typography/HeadingDeprecated';
-export * from './Typography/Text';
-export * from './Typography/TextDeprecated';
 export * from './HighlightedText';
 export * from './InputStepper';
-export * from './Interstitial';
 export * from './Layout/Column';
 export * from './Layout/LayoutGrid';
+export * from './List';
+export * from './Loading/Shimmer';
+export * from './Logo';
 export * from './Markdown';
-export * from './Modals';
+export * from './Menu';
 export * from './ModalDeprecated';
+export * from './Modals';
 export * from './NotificationList';
 export * from './NotificationList/NotificationItem';
 export * from './NotificationList/typings';
 export * from './Overlay';
 export * from './Pagination';
+export * from './Popover';
+export * from './PopoverContainer';
 export * from './ProgressBar';
+export * from './ProLabel';
 export * from './RadialProgress';
 export * from './SkipToContent';
 export * from './Spinner';
-export * from './theming/VisualTheme';
 export * from './Tabs';
+export * from './theming/VisualTheme';
+export * from './Toast/Toast';
+export * from './Toaster';
 export * from './Toggle';
 export * from './ToolTip';
 export * from './Truncate';
+export * from './typings/responsive-properties';
+export * from './Typography/HeadingDeprecated';
+export * from './Typography/Text';
+export * from './Typography/TextDeprecated';
 export * from './utils';
 export * from './Video';
-export * from './typings/responsive-properties';
-export * from './Logo';
-export * from './ProLabel';
-export * from './PopoverContainer';
-export * from './Toast/Toast';
-export * from './Toaster';
-export * from './FloatingCard/FloatingCard';
-export * from './BetaSticker';
-export * from './Menu';
-export * from './List';
-export * from './DataList';
-export * from './Loading/Shimmer';
