+++ packages/styleguide/stories/Brand Labs/Molecules/PausableImage.stories.mdx
